+++ packages/react-dom/src/client/ReactDOMHostConfig.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-import type {TopLevelType} from '../legacy-events/TopLevelEventTypes';
+import type {TopLevelType} from '../events/TopLevelEventTypes';
 import type {Fiber, FiberRoot} from 'react-reconciler/src/ReactInternalTypes';
 import type {
   BoundingRect,
