+++ packages/react-dom/src/events/plugins/ModernEnterLeaveEventPlugin.js
@@ -145,7 +145,7 @@ const EnterLeaveEventPlugin = {
     const fromNode = from == null ? win : getNodeFromInstance(from);
     const toNode = to == null ? win : getNodeFromInstance(to);
 
-    const leave = eventInterface.getPooled(
+    const leave = new eventInterface(
       leaveEventType,
       from,
       nativeEvent,
@@ -155,7 +155,7 @@ const EnterLeaveEventPlugin = {
     leave.target = fromNode;
     leave.relatedTarget = toNode;
 
-    let enter = eventInterface.getPooled(
+    let enter = new eventInterface(
       enterEventType,
       to,
       nativeEvent,
