+++ packages/react-dom/src/test-utils/ReactTestUtilsAct.js
@@ -20,11 +20,9 @@ const [
   getInstanceFromNode,
   getNodeFromInstance,
   getFiberCurrentPropsFromNode,
-  injectEventPluginsByName,
   eventNameDispatchConfigs,
   enqueueStateRestore,
   restoreStateIfNeeded,
-  dispatchEvent,
   /* eslint-enable no-unused-vars */
   flushPassiveEffects,
   IsThisRendererActing,
